feat: moves Homescreen to homepage path
