--- src/App.tsx
@@ -1,7 +1,0 @@
-"use client";
-
-import { MultiStepForm } from "@/components";
-
-export default function App() {
-	return <MultiStepForm />;
-}
